deen: audio: Correct tfa98xx_adsp_send_calib_values to mono mode
-Becasue the smartpa of deen is mono, so need remove the stereo condition
and add mono condition to ensure dsp can use actual calibration resistance
of smartpa,but not the Defenderault value in dsp.
Now it is wrong status and there is any problem,because the Defenderault value
is close to actual value of calibration resistance.So it is corrected
immediately.

Change-Id: I0b7acf491a17d4fb2360aa1ee20537f8e70408ab
Signed-off-by: lily8 <lily8@lenovo.com>
Reviewed-on: https://gerrit.mot.com/1227170
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Wanlong Zhou <zhouwl@motorola.com>
Reviewed-by: Hongshu Lou <louhs1@motorola.com>
Submit-Approved: Jira Key